fix(lens): App.tsx restart handler owns multi-step RPC workflow against SolidJS convention
Extracted handleRestartPtyHost into a new useDaemonRestart.ts singleton; App.tsx now wires daemonRestart.restartPtyHost into the command deps and no longer owns the workflow.

Agreed by the lowy ⇄ hickey lens debate (finding hickey-7, raised by hickey). Not pushed or merged.
